requirements.txt
ultralytics
flask
gunicorn
Pillow